+++ src/Entities/Transaction.php
@@ -63,6 +63,7 @@ class Transaction implements Entity
         public Customer|null $customer,
         public Discount|null $discount,
         public array $availablePaymentMethods,
+        public \DateTimeInterface|null $revisedAt,
     ) {
     }
 
@@ -98,6 +99,7 @@ class Transaction implements Entity
             customer: isset($data['customer']) ? Customer::from($data['customer']) : null,
             discount: isset($data['discount']) ? Discount::from($data['discount']) : null,
             availablePaymentMethods: array_map(fn (string $item): AvailablePaymentMethods => AvailablePaymentMethods::from($item), $data['available_payment_methods'] ?? []),
+            revisedAt: isset($data['revised_at']) ? DateTime::from($data['revised_at']) : null,
         );
     }
 }