+++ tests/Functional/Resources/Notifications/NotificationsClientTest.php
@@ -7,11 +7,14 @@ namespace Paddle\SDK\Tests\Functional\Resources\Notifications;
 use GuzzleHttp\Psr7\Response;
 use Http\Mock\Client as MockClient;
 use Paddle\SDK\Client;
+use Paddle\SDK\Entities\DateTime;
 use Paddle\SDK\Entities\Notification;
 use Paddle\SDK\Entities\Notification\NotificationStatus;
 use Paddle\SDK\Environment;
 use Paddle\SDK\Notifications\Entities\Entity;
+use Paddle\SDK\Notifications\Entities\Transaction;
 use Paddle\SDK\Notifications\Entities\UndefinedEntity;
+use Paddle\SDK\Notifications\Events\TransactionRevised;
 use Paddle\SDK\Notifications\Events\UndefinedEvent;
 use Paddle\SDK\Options;
 use Paddle\SDK\Resources\Notifications\Operations\ListNotifications;
@@ -209,6 +212,36 @@ class NotificationsClientTest extends TestCase
                 'key' => 'value',
             ],
             $undefinedEvent->entity->data,
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function list_transaction_revised_notification(): void
+    {
+        $this->mockClient->addResponse(new Response(200, body: self::readRawJsonFixture('response/list_default')));
+        $notifications = $this->client->notifications->list(new ListNotifications());
+
+        $transactionRevisedNotifications = array_values(
+            array_filter(
+                iterator_to_array($notifications),
+                fn (Notification $notification) => (string) $notification->type === 'transaction.revised',
+            ),
+        );
+
+        $transactionRevisedNotification = $transactionRevisedNotifications[0];
+        self::assertInstanceOf(Notification::class, $transactionRevisedNotification);
+
+        $transactionRevisedEvent = $transactionRevisedNotification->payload;
+        self::assertInstanceOf(TransactionRevised::class, $transactionRevisedEvent);
+        self::assertSame($transactionRevisedEvent->transaction, $transactionRevisedEvent->data);
+        self::assertInstanceOf(Entity::class, $transactionRevisedEvent->data);
+        self::assertInstanceOf(Transaction::class, $transactionRevisedEvent->data);
+        self::assertInstanceOf(Transaction::class, $transactionRevisedEvent->transaction);
+        self::assertEquals(
+            '2024-04-12T10:18:50.738972Z',
+            $transactionRevisedEvent->transaction->revisedAt->format(DateTime::PADDLE_RFC3339),
         );
     }
 }